+++ CHANGELOG.md
@@ -3,6 +3,20 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.7.0](https://github.com/martijnvdbrug/pinelab-vendure-plugins/compare/v4.6.0...v4.7.0) (2022-12-19)
+
+### Features
+
+- lerna git stash before publish ([ac175a7](https://github.com/martijnvdbrug/pinelab-vendure-plugins/commit/ac175a7ba66367942d77cb96abe90d1d4c024544))
+- **limit-variant:** also patches adjustOrderLine ([7199a1f](https://github.com/martijnvdbrug/pinelab-vendure-plugins/commit/7199a1f94593448304b929fe9024e851d7cdc749))
+
+## [4.6.0](https://github.com/martijnvdbrug/pinelab-vendure-plugins/compare/v4.5.0...v4.6.0) (2022-12-19)
+
+### Features
+
+- **limit-variant:** finished plugin ([1a370cb](https://github.com/martijnvdbrug/pinelab-vendure-plugins/commit/1a370cb730a2dbbb8925d300b85ee54731d98f45))
+- **limit-variant:** package main file ([e6a7450](https://github.com/martijnvdbrug/pinelab-vendure-plugins/commit/e6a7450781cad639409f9fc20d0c5c6adbcdc953))
+
 ## [4.5.0](https://github.com/martijnvdbrug/pinelab-vendure-plugins/compare/v4.4.0...v4.5.0) (2022-12-14)
 
 ### Features